requirements.txt
streamlit
openai
pandas
pydantic
numpy
python-dotenv
